Remove classes moved to pyxavi-activitypub and bump version
