+++ tests/test_storage.py
@@ -1,5 +1,5 @@
 from unittest.mock import patch, mock_open
-from pyxavi.storage import Storage
+from pyxavi import Storage
 from unittest import TestCase
 import pytest
